+++ ChatMacros/_Thunderstore/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-Dec-05
+
+### Added
+
+- Alt and Ctrl Macro combinations can now be used, raising the total number from 9 to 27
+- Added `MacroNAlt` options to configure Alt behaviour for macros
+- Added `MacroNCtrl` options to configure Ctrl behaviour for macros
+- Added `EnableAltMacros` and `EnableCtrlMacros` options to control whenever Alt and Ctrl macros are active
+
 ## [1.0.1] - 2025-Dec-04
 
 ### Fixed